Move resolveId logic
